[FIX] absi- _transfer_move. Correct test and duplicate class name
