+++ account_bank_statement_import_transfer_move/models/account_journal.py
@@ -4,7 +4,7 @@
 from odoo import fields, models
 
 
-class AccountBankStatementImport(models.Model):
+class AccountJournal(models.Model):
 
     _inherit = "account.journal"
